Add files via upload
The Scratch project is a simple basketball game. The user clicks on the basketball to try to shoot it into the hoop. There is a dinosaur that moves around that I was considering having attempt to "block" shots if the ball crossed its path, but for now it is just there for show.